pyproject.toml
[project]
name = "aac"
version = "0.1.0"
description = "Agent Application Context — Spring-inspired IoC/DI/AOP framework for AI Agent orchestration"
requires-python = ">=3.12"
dependencies = [
    "fastapi>=0.115.0",
    "uvicorn[standard]>=0.32.0",
    "pyyaml>=6.0.2",
    "pydantic>=2.10.0",
    "click>=8.1.0",
    "structlog>=24.4.0",
    "rich>=13.9.0",
]

[project.optional-dependencies]
tui = ["textual>=1.0.0"]
dev = [
    "pytest>=8.3.0",
    "pytest-asyncio>=0.24.0",
    "httpx>=0.27.0",
    "ruff>=0.8.0",
]
all = ["aac[tui,dev]"]

[project.scripts]
aac = "aac.cli.main:cli"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/aac", "src/ui"]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"
pythonpath = ["src"]

[tool.ruff]
target-version = "py312"
line-length = 100
src = ["src", "tests"]

[tool.ruff.lint]
select = ["E", "F", "I", "N", "W", "UP"]
